requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
bcrypt==4.2.1
PyJWT==2.10.0
requests==2.32.3
pandas==2.2.3
geopy==2.4.1
fuzzywuzzy==0.18.0
